Updated interactors

+++ fb_post_v2/interactors/delete_post_interactor.py
@@ -8,9 +8,10 @@ class DeletePostInteractor:
         self.post_storage = post_storage
         self.presenter = presenter
 
-    def create_post(self, post_id: int):
-        status_dict = self.post_storage.delete_post(post_id)
-        response = self.presenter.delete_post(status_dict)
-        return response
-
-
+    def delete_post(self, post_id: int):
+        post_exists = self.post_storage.post_exists(post_id)
+        if post_exists:
+            status_dict = self.post_storage.delete_post(post_id)
+            response = self.presenter.delete_post(status_dict)
+            return response
+        return self.presenter.raise_post_does_not_exist_exception()
